Create new dedicated components for ProjectCard
